Part of https://github.com/Gamify-IT/issues/issues/498

https://sonar.sqa-kube-cluster4.cloudns.nz/dashboard?id=memory-backend

**Please test is everything is still working as intendet**

+++ src/main/java/de/unistuttgart/memorybackend/service/GameResultService.java
@@ -72,6 +72,6 @@ public class GameResultService {
     }
 
     private int calculateResultScore(final boolean isCompleted) {
-        return (int) (isCompleted ? 100 : 0);
+        return (isCompleted ? 100 : 0);
     }
 }
